4.Flask/4.2-10.Flask/requirements.txt
flask
bcrypt
sqlmodel
pillow
numpy
opencv-python
onnxruntime
pyyaml
psycopg2
scikit-image
